refactor(workers): debounce repeat sync failures and extract the scheduler job registry
A 5-minute repeat job that rethrows on an unreachable Audiobookshelf
logs two error lines per cycle (~576/day). Repeat-mode failures are now
caught: warn at most once per hour, debug otherwise, and the job
resolves so Bull does not record a failure storm. Startup-mode failures
still propagate unchanged.

The added handling pushed workers/index.ts over the 1,500-line ratchet,
so the scheduler job registry (job types, job IDs, registration
entries, shared time constants) moves verbatim to
schedulerJobRegistry.ts — the established remedy of extracting a
cohesive block instead of squeezing lines. buildSchedulerJobs() is
called at registration time, preserving per-boot sweepStartedAt
timestamps. index.ts drops to 1,230 lines.

New runtime coverage: repeat failures resolve and warn, the warning
debounces to debug within the hour window, and startup failures still
reject.
